% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bi_tsls.R
\name{bi_tsls_estimation}
\alias{bi_tsls_estimation}
\title{Bi-TSLS Estimation Algorithm}
\usage{
bi_tsls_estimation(data, R_w, R_z)
}
\arguments{
\item{data}{Prepared data frame with standardized variables}

\item{R_w}{Sensitivity parameter R_w}

\item{R_z}{Sensitivity parameter R_z}
}
\value{
Vector of estimated beta_xy and beta_yx
}
\description{
Bi-TSLS Estimation Algorithm
}
\keyword{internal}
