% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{compare_algorithm_with_r}
\alias{compare_algorithm_with_r}
\title{Compare algorithm with different R values}
\usage{
compare_algorithm_with_r(
  data,
  x_var,
  y_var,
  z_var,
  w_var,
  covariates = NULL,
  start_value = -0.5,
  step_size = 0.1,
  num_steps = 10,
  num_iterations = 500,
  parallel = TRUE,
  num_cores = NULL
)
}
\arguments{
\item{data}{Dataset to use for simulation}

\item{x_var}{Name of X variable}

\item{y_var}{Name of Y variable}

\item{z_var}{Name of Z variable}

\item{w_var}{Name of W variable}

\item{covariates}{Optional vector of covariate variable names}

\item{start_value}{Starting value for R_w (R_z will be the negative)}

\item{step_size}{Step size for R value changes}

\item{num_steps}{Number of steps to take}

\item{num_iterations}{Number of bootstrap iterations per R combination}

\item{parallel}{Whether to use parallel processing (default: TRUE)}

\item{num_cores}{Number of cores to use for parallel processing (default: NULL, uses detectCores())}
}
\value{
List of results for each R combination
}
\description{
Compare algorithm with different R values
}
\examples{
\dontrun{
data <- read.csv("your_data.csv")
results <- compare_algorithm_with_r(data, "X_var", "Y_var", "Z_var", "W_var",
                                   start_value = -0.5,
                                   step_size = 0.1, num_steps = 10)
}
}
