% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{generate_r_combinations}
\alias{generate_r_combinations}
\title{Generate R value combinations}
\usage{
generate_r_combinations(start_value, step_size, num_steps)
}
\arguments{
\item{start_value}{Starting value for R_w (R_z will be the negative)}

\item{step_size}{Step size for R value changes}

\item{num_steps}{Number of steps to take}
}
\value{
List of R_w and R_z combinations
}
\description{
Generate R value combinations
}
\examples{
generate_r_combinations(-0.5, 0.1, 10)
}
