% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_bi_tsls_strength_YX}
\alias{plot_bi_tsls_strength_YX}
\title{Plot Y to X causal effect across different R values}
\usage{
plot_bi_tsls_strength_YX(
  df_cleaned,
  save_plot = FALSE,
  filename = "Bi_TSLS_Strength_YX_lines.png",
  width = 10,
  height = 6
)
}
\arguments{
\item{df_cleaned}{Cleaned data frame from process_simulation_results}

\item{save_plot}{Whether to save the plot to a file (default: FALSE)}

\item{filename}{Filename to save plot (default: "Bi_TSLS_Strength_YX_lines.png")}

\item{width}{Plot width in inches (default: 10)}

\item{height}{Plot height in inches (default: 6)}
}
\value{
ggplot object
}
\description{
Plot Y to X causal effect across different R values
}
