% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bi_tsls.R
\name{prepare_data}
\alias{prepare_data}
\title{Prepare data for Bi-TSLS analysis}
\usage{
prepare_data(
  data,
  x_var,
  y_var,
  z_var,
  w_var,
  covariates = NULL,
  standardize_vars = TRUE
)
}
\arguments{
\item{data}{Original data frame}

\item{x_var}{Name of X variable}

\item{y_var}{Name of Y variable}

\item{z_var}{Name of Z variable}

\item{w_var}{Name of W variable}

\item{covariates}{Optional vector of covariate variable names}

\item{standardize_vars}{Logical, whether to standardize Z and covariate variables}
}
\value{
Prepared data frame
}
\description{
Prepare data for Bi-TSLS analysis
}
\keyword{internal}
