% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{process_simulation_results}
\alias{process_simulation_results}
\title{Process simulation results into a data frame}
\usage{
process_simulation_results(results, R_combinations)
}
\arguments{
\item{results}{List of results from compare_algorithm_with_r}

\item{R_combinations}{List of R combinations used}
}
\value{
Data frame with processed results
}
\description{
Process simulation results into a data frame
}
