% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{remove_extreme_outliers}
\alias{remove_extreme_outliers}
\title{Remove extreme outliers from a vector}
\usage{
remove_extreme_outliers(x, probs = c(0.05, 0.95))
}
\arguments{
\item{x}{Numeric vector}

\item{probs}{Probability quantiles to use as cutoffs (default: c(0.05, 0.95))}
}
\value{
Vector with extreme values replaced by NA
}
\description{
Remove extreme outliers from a vector
}
\keyword{internal}
